requirements.txt
altair
pandas
streamlit
google-genai
faiss-cpu
sentence_transformers
tqdm
pandas
numpy
tiktoken
langchain
pypdf
zai-sdk